refactor : 의존성 관계 수정
