+++ app/src/main/java/com/example/system/IngredientScreen.kt
@@ -1,5 +1,6 @@
 package com.example.system
 
+import android.content.Context
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -12,14 +13,11 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.Button
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableLongStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
@@ -30,7 +28,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.example.system.ingredientsDB.Ingredient
-import com.google.android.libraries.places.api.model.kotlin.localDate
+import com.example.system.ui.viewmodel.IngredientViewModel
 import java.time.LocalDate
 import java.time.ZoneId
 
@@ -94,7 +92,8 @@ fun init(ingredientViewModel: IngredientViewModel) {
 
 @Composable
 fun IngredientScreen(
-    ingredientViewModel: IngredientViewModel
+    context : Context,
+    ingredientViewModel: IngredientViewModel = IngredientViewModel(context)
 ) {
     init(ingredientViewModel)
