+++ app/src/main/java/com/example/system/ingredientsDB/IngredientDB.kt
@@ -5,6 +5,7 @@ import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
+import dagger.hilt.android.qualifiers.ApplicationContext
 
 @Database(entities = [Ingredient::class], version = 1)
 abstract class IngredientDB : RoomDatabase() {
